% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_pileupId.R
\name{build_pileupId}
\alias{build_pileupId}
\title{Build coverage pileup based on annotations using NewSampleIds (modified build_pileup)}
\usage{
build_pileupId(
  Pileup,
  caseIDs = NULL,
  regions,
  inputType = "part_intron",
  outputType = "part_intron"
)
}
\description{
Build coverage pileup based on annotations using NewSampleIds (modified build_pileup)
}
\references{
https://github.com/hyochoi/SCISSOR
}
