% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_Chr2Numcol.R
\name{convert_Chr2Numcol}
\alias{convert_Chr2Numcol}
\title{Convert character columns to numeric columns in a data frame}
\usage{
convert_Chr2Numcol(df, colnums)
}
\arguments{
\item{df}{a data frame}

\item{colnums}{the column numbers to be converted}
}
\description{
Convert character columns to numeric columns in a data frame
}
\references{
https://www.geeksforgeeks.org/how-to-convert-dataframe-column-from-character-to-numeric-in-r/#
}
