% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_Cont2eCDF.gene.R
\name{convert_Cont2eCDF.gene}
\alias{convert_Cont2eCDF.gene}
\title{Convert a continuous variable to eCDF values from ecdf function (For gene information)}
\usage{
convert_Cont2eCDF.gene(data, name)
}
\arguments{
\item{data}{a continuous variable for x-axis in the eCDF plot; a 1 x the number of genes matrix; rownames are geneSymbol}

\item{name}{a column name for eCDF Values (name_ecdf)}
}
\description{
Convert a continuous variable to eCDF values from ecdf function (For gene information)
}
\references{
https://statisticsglobe.com/extract-ecdf-values-from-function-r

https://stats.stackexchange.com/questions/30858/how-to-calculate-cumulative-distribution-in-r
}
