% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_pivot.longer.R
\name{convert_pivot.longer}
\alias{convert_pivot.longer}
\title{Convert a matrix to a data frame using pivot_longer}
\usage{
convert_pivot.longer(mat, rcenames)
}
\arguments{
\item{mat}{a matrix with observations in rows, variables in columns, and values in cells}

\item{rcenames}{row, column, and cell names of a matrix}
}
\description{
Convert a matrix to a data frame using pivot_longer
}
\references{
https://dcl-wrangle.stanford.edu/pivot-basic.html
}
