% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_metrics.R
\name{get_metrics}
\alias{get_metrics}
\title{Get metrics from scaled normalized transcript coverage}
\usage{
get_metrics(
  pileupPath,
  geneNames = NULL,
  rnum = 100,
  method = 1,
  scale = TRUE,
  margin
)
}
\arguments{
\item{pileupPath}{file paths of coverage pileupData including .RData file names}

\item{geneNames}{gene names per file. If NULL, Gene i with the same length of pileupPath be set. Default is NULL.}

\item{rnum}{the number of regions for uniformly dividing the x-axis. Default is 100.}

\item{method}{1 and 2 return the raw read depth and the interpolated read depth at the normalized genomic position, respectively. Default is 1.}

\item{scale}{TRUE/FALSE returns the scaled/unscaled normalized transcript coverage. Default is TRUE.}

\item{margin}{1, 2, and 3 return metrics per sample, per gene, and across the genes per sample, respectively.}
}
\value{
metrics including mean, sd, CV (sd/mean), median, mad, and robustCV (mad/median) per margin
}
\description{
Get metrics from scaled normalized transcript coverage
}
\references{
https://github.com/hyochoi/SCISSOR
}
