% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm_pileup.list.R
\name{norm_pileup.list}
\alias{norm_pileup.list}
\title{Gene length normalization for pileup lists (for all genes, all samples)}
\usage{
norm_pileup.list(pileupPath, geneNames = NULL, rnum = 100, method = 1)
}
\arguments{
\item{pileupPath}{file paths of coverage pileupData including .RData file names}

\item{geneNames}{gene names per file. If NULL, Gene i with the same length of pileupPath be set. Default is NULL.}

\item{rnum}{the number of regions for uniformly dividing the x-axis. Default is 100.}

\item{method}{1 and 2 return the raw read depth and the interpolated read depth at the normalized genomic position, respectively. Default is 1.}
}
\value{
the normalized read depth is a rnum x the number of samples matrix at each gene list.
}
\description{
Gene length normalization for pileup lists (for all genes, all samples)
}
\references{
https://github.com/hyochoi/SCISSOR
}
