% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm_pileup.spl.R
\name{norm_pileup.spl}
\alias{norm_pileup.spl}
\title{Gene length normalization for a single pileup (for gene 1, sample 1)}
\usage{
norm_pileup.spl(pileup, rnum = 100, method = 1)
}
\arguments{
\item{pileup}{a coverage pileup vector for one sample}

\item{rnum}{the number of regions for uniformly dividing the x-axis. Default is 100.}

\item{method}{1 and 2 return the raw read depth and the interpolated read depth at the normalized genomic position, respectively. Default is 1.}
}
\value{
the normalized read depth is a vector with length=rnum.
}
\description{
Gene length normalization for a single pileup (for gene 1, sample 1)
}
\references{
https://github.com/hyochoi/SCISSOR
}
