% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pileupPair.R
\name{plot_pileupPair}
\alias{plot_pileupPair}
\title{Plot pileup profiles with three pair lines (modified plot_pileup)}
\usage{
plot_pileupPair(
  AllPileup,
  Pileup,
  Ranges,
  cases = NULL,
  logcount = NULL,
  plot.meanpileup = TRUE,
  col.pileup = NULL,
  col.meanpileup = "grey",
  main = NULL,
  cex.main = 1.2,
  print.ranges = TRUE,
  xlim = NULL,
  ylim = NULL,
  xlab = NULL,
  ylab = NULL,
  ...
)
}
\arguments{
\item{AllPileup}{pileup for all samples to make the same y-axis with other samples}

\item{Pileup}{pileup for selected samples from AllPileup}
}
\description{
Plot pileup profiles with three pair lines (modified plot_pileup)
}
\references{
https://github.com/hyochoi/SCISSOR
}
