% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pileupStat.R
\name{plot_pileupStat}
\alias{plot_pileupStat}
\title{Plot median or mean pileup profiles within each group (modified plot_pileup)}
\usage{
plot_pileupStat(
  AllPileup,
  Pileup,
  Ranges,
  cases = NULL,
  logcount = NULL,
  plot.meanpileup = TRUE,
  col.pileup = NULL,
  col.meanpileup = "grey",
  main = NULL,
  cex.main = 1.2,
  print.ranges = TRUE,
  xlim = NULL,
  ylim = NULL,
  xlab = NULL,
  ylab = NULL,
  pair,
  ...
)
}
\arguments{
\item{AllPileup}{pileup for all samples to make the same y-axis with other samples}

\item{Pileup}{pileup for selected samples from AllPileup}

\item{pair}{FFM or FFT or PET to set a different color line by pair}
}
\description{
Plot median or mean pileup profiles within each group (modified plot_pileup)
}
\references{
https://github.com/hyochoi/SCISSOR
}
