% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repeach.R
\name{repeach}
\alias{repeach}
\title{Repeat rows of a matrix}
\usage{
repeach(mat, n)
}
\arguments{
\item{mat}{a matrix}

\item{n}{the number of repeat rows}
}
\description{
Repeat rows of a matrix
}
\references{
https://stackoverflow.com/questions/11121385/repeat-rows-of-a-data-frame
}
