% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repmat.R
\name{repmat}
\alias{repmat}
\title{Repeat copies of array (equivalent of repmat in Matlab)}
\usage{
repmat(X, m, n)
}
\arguments{
\item{X}{a matrix}

\item{m}{m copies of X are arranged in row}

\item{n}{n copies of X are arranged in column}
}
\description{
Repeat copies of array (equivalent of repmat in Matlab)
}
\references{
Hua Liu, Jinhong You & Jiguo Cao (2023). A Dynamic Interaction Semiparametric Function-on-Scalar Model, Journal of the American Statistical Association, 118:541, 360-373, DOI: 10.1080/01621459.2021.1933496
}
