% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpl_NewSampleId.R
\name{rpl_NewSampleId}
\alias{rpl_NewSampleId}
\title{Replace colnames of a matrix based on NewSampleId in Pilot2}
\usage{
rpl_NewSampleId(mat, Pilot2)
}
\arguments{
\item{mat}{a matrix with entity_submitter_id columns}

\item{Pilot2}{a reference matrix having NewSampleId}
}
\description{
Replace colnames of a matrix based on NewSampleId in Pilot2
}
