% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_pileup.list.R
\name{scale_pileup.list}
\alias{scale_pileup.list}
\title{Scale normalized transcript coverage using gene length normalization}
\usage{
scale_pileup.list(
  pileupPath,
  geneNames = NULL,
  rnum = 100,
  method = 1,
  scale = TRUE
)
}
\arguments{
\item{pileupPath}{file paths of coverage pileupData including .RData file names}

\item{geneNames}{gene names per file. If NULL, Gene i with the same length of pileupPath be set. Default is NULL.}

\item{rnum}{the number of regions for uniformly dividing the x-axis. Default is 100.}

\item{method}{1 and 2 return the raw read depth and the interpolated read depth at the normalized genomic position, respectively. Default is 1.}

\item{scale}{TRUE/FALSE returns the scaled/unscaled normalized transcript coverage. Default is TRUE.}
}
\value{
gene lists (rows are regions and columns are samples) for the normalized transcript coverage after gene length normalization
}
\description{
Scale normalized transcript coverage using gene length normalization
}
\references{
https://github.com/hyochoi/SCISSOR
}
